refactor: exception global로 처리
